+++ NEWS.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.15.5] 25/11/2023
+
+### Added
+
+* An `AbstractEstimationMethod` to specify estimation methods for other more general functions,
+as well as a `get_default_estimation_method` to specify defaults on manifolds.
+
 ## [0.15.4] 25/11/2023
 
 ### Fixed